+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.20.0 - (Spring Green Vienna) - June 2022
+
+- added default value (`aggregates.mdsp.json`) for `mdsp aggregates --download` parameter
+- improved `mdsp aggregates` command. all parameters, including intervalunit and intervalvalue can be specified.
+- Created a new `mdsp timeseries` command to match the `mdsp aggregates` command. [#318](https://github.com/mindsphere/mindconnect-nodejs/issues/318)
+- Bumped most depedencies
+
 ## 3.19.0 - (Dove Vienna) - April 2022
 
 - CLI: new `mdsp markdown-help` command to generate markdown help available at [opensource.mindsphere.io](https://opensource.mindsphere.io/docs/mindconnect-nodejs/cli/index.html#tab1anchor2)